requirements.txt
scikit-learn==1.5.1
numpy==1.26.4
PyYAML==6.0.2
